refactor: Split openapiSpec.ts into multiple files
