+++ src/utils/strings.ts
@@ -1,0 +1,2 @@
+export const upperFirst = (value: string) => value[0].toUpperCase() + value.slice(1)
+export const camelize = (value: string) => value.split(/\s+/).map(upperFirst).join('')
